requirements.txt
python-dotenv
PyMySQL